# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Jersey {
  id: number;
  name: string;
  team: string;
  country: string;
  /** home, away, or third */
  type: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  inStock: boolean;
  sizes: string[];
  isFeatured?: boolean;
  rating?: number;
  reviewCount?: number;
}

export interface CartItem {
  jerseyId: number;
  quantity: number;
  size: string;
}

export interface OrderInput {
  items: CartItem[];
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  city?: string;
  postalCode?: string;
  country?: string;
  /** Verified OTP token received after OTP verification */
  otpToken: string;
}

export interface Order {
  id: string;
  /** pending, confirmed, shipped, delivered */
  status: string;
  customerName: string;
  customerEmail: string;
  shippingAddress?: string;
  totalAmount: number;
  createdAt: string;
}

export interface AdminOrder {
  id: string;
  status: string;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  country?: string | null;
  totalAmount: number;
  items: CartItem[];
  createdAt: string;
}

export interface AdminOrderList {
  orders: AdminOrder[];
  total: number;
  limit?: number;
  offset?: number;
}

export type UpdateOrderStatusInputStatus = typeof UpdateOrderStatusInputStatus[keyof typeof UpdateOrderStatusInputStatus];


export const UpdateOrderStatusInputStatus = {
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export type AdminStatsOrdersByStatusItem = {
  status?: string;
  count?: number;
};

export type AdminStatsRevenueByDayItem = {
  date?: string;
  revenue?: number;
  orders?: number;
};

export type AdminStatsTopTeamsItem = {
  team?: string;
  units?: number;
  revenue?: number;
};

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  avgOrderValue?: number;
  ordersByStatus: AdminStatsOrdersByStatusItem[];
  revenueByDay: AdminStatsRevenueByDayItem[];
  topTeams: AdminStatsTopTeamsItem[];
}

export interface OtpRequest {
  email: string;
}

export interface OtpRequestResult {
  message: string;
  /** OTP shown in demo mode (remove in production) */
  devOtp: string;
}

export interface OtpVerify {
  email: string;
  otp: string;
}

export interface OtpVerifyResult {
  message: string;
  token: string;
}

export interface PaymentIntentInput {
  /** Amount in smallest currency unit (e.g. cents for USD) */
  amountCents: number;
  /** ISO 4217 currency code (e.g. usd, gbp) */
  currency: string;
}

export interface PaymentIntentResult {
  clientSecret: string;
  paymentIntentId: string;
}

export type ListJerseysParams = {
/**
 * Filter by team name
 */
team?: string;
/**
 * Filter by type (home/away/third)
 */
type?: string;
};

export type ListAdminOrdersParams = {
/**
 * Filter by status (confirmed, shipped, delivered, cancelled)
 */
status?: string;
/**
 * Search by customer name or email
 */
search?: string;
limit?: number;
offset?: number;
};