# lib/api-zod/src/generated/types/adminOrderList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminOrder } from './adminOrder';

export interface AdminOrderList {
  orders: AdminOrder[];
  total: number;
  limit?: number;
  offset?: number;
}