# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsOrdersByStatusItem } from './adminStatsOrdersByStatusItem';
import type { AdminStatsRevenueByDayItem } from './adminStatsRevenueByDayItem';
import type { AdminStatsTopTeamsItem } from './adminStatsTopTeamsItem';

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  avgOrderValue?: number;
  ordersByStatus: AdminStatsOrdersByStatusItem[];
  revenueByDay: AdminStatsRevenueByDayItem[];
  topTeams: AdminStatsTopTeamsItem[];
}