# lib/api-zod/src/generated/types/jersey.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Jersey {
  id: number;
  name: string;
  team: string;
  country: string;
  /** home, away, or third */
  type: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  inStock: boolean;
  sizes: string[];
  isFeatured?: boolean;
  rating?: number;
  reviewCount?: number;
}