# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface OrderInput {
  items: CartItem[];
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  city?: string;
  postalCode?: string;
  country?: string;
  /** Verified OTP token received after OTP verification */
  otpToken: string;
}