# lib/api-zod/src/generated/types/otpRequestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpRequestResult {
  message: string;
  /** OTP shown in demo mode (remove in production) */
  devOtp: string;
}