# lib/api-zod/src/generated/types/otpVerifyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpVerifyResult {
  message: string;
  token: string;
}