# lib/api-zod/src/generated/types/paymentIntentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saifit — FIFA Jersey Store API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentIntentInput {
  /** Amount in smallest currency unit (e.g. cents for USD) */
  amountCents: number;
  /** ISO 4217 currency code (e.g. usd, gbp) */
  currency: string;
}